Create Demonstration Code
The purpose of the project is to demonstrate the use of git submodules. This code will be used to demonstrate how one may switch between branched of the a submodule in order to explore changes to the functions of the submodule, as they are in development, within the context of another project where they may be applied.